Refactor geo configuration to separate config.go file
