+++ geo/geopoint.go
@@ -8,23 +8,9 @@ import (
 var polygon = "Polygon"
 
 const (
-	twoPi                  float64 = 2 * math.Pi
-	radiusOfEarth          float64 = 6378137 //(in metres defined by wgs84)
-	defaultConcurrency             = 10      // limit number of go routines to not put too much on heap
-	defaultMaximumSegments         = 180
+	twoPi         float64 = 2 * math.Pi
+	radiusOfEarth float64 = 6378137 //(in metres defined by wgs84)
 )
-
-// Config object to define geo configurations
-type Config struct {
-	defaultConcurrencyLimit int
-	defaultMaxSegments      int
-}
-
-// Default geo configuration for methods on config receiver
-var DefaultConfig = &Config{
-	defaultConcurrencyLimit: defaultConcurrency,
-	defaultMaxSegments:      defaultMaximumSegments,
-}
 
 // GeoStructure describes the shape of the geographical location
 type GeoStructure struct {
@@ -36,17 +22,6 @@ type GeoStructure struct {
 type Coordinate struct {
 	Lat float64
 	Lon float64
-}
-
-// New instantiates a geo object with defined values to be used by geo methods
-// e.g. concurrency limit and maximum number of segments when generating a circle from a point
-func New(concurrencyLimit, maximumSegments int) (geo *Config) {
-	geo = DefaultConfig
-
-	geo.defaultConcurrencyLimit = concurrencyLimit
-	geo.defaultMaxSegments = maximumSegments
-
-	return
 }
 
 // NewGeoPoint creates a coordinate object
